updated generate() function
updated this function to use the built-in secrets library to generate a "true" random, url-safe string, 20 chars in length, as opposed to my home-baked, pseudo-random solution from the Udemy course project, which used the random library. This is in the interest of password security, as the secrets library uses SystemRandom to generate random numbers that are much harder to guess than the random library's random functions.