added encryption master pass
Added sha256 salted encryption layer to the master password verification layer.

The master password is no longer stored in plaintext. The salt is stored in the salt.key file, and the encrypted, salted password is stored in the master.key file.

When the program runs, it checks for the existence of the salt.key and master.key files, and if they don't exist, they're created.

When the a password is requested and found in the data file, a popup window requests the master password. 
This password is salted, and then encrypted and checked against the master.key file in storage. 

If there's a match, the password data is displayed in a subsequent popup window. 

The next update will add an encryption layer to the stored password data in our .json file.