updated username/email input
No longer hard-coded in. Before we set a master password, a prompt appears to set a default username/email. This is saved into a plaintext file. The default username field (user@example.com) is deleted from the text box, and replaced with our preferred default username/email. This can be changed for individual logins, but will auto-populate upon running a fresh instance of the program.